Reverse proxy updates.
When running in server mode with a reverse-proxy defined, in the event that
a proxied request for a non-managed resource returns 404, the server now
attempts to locate the resource locally instead of just forwarding the 404
error to the client.

This mirrors the handling for managed resources. If those cannot be found
locally the request is forwarded to the proxy.