Route mapping improvements
Routes mapping now allows a single path to match more than one target. For example,
the following route:

	/assets = [ "/public", "/private" ]

will expand to check both targets in the order they are defined for paths matching
the left side of the mapping. In this case, the input path:

	/assets/css/style.css

would produce the candidates:

	/public/css/style.scss
	/private/css/style.scss
	/public/css/style.css
	/private/css/style.css

One-to-one route mapping is still supported. The following is still valid:

	/assets = "/public"